pyproject.toml
[project]
name = "extractosm"
version = "0.1.0"
description = "Convenience package to extract various features from OpenStreetMap"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geopandas>=1.1.3",
    "osmium>=4.3.0",
    "osmnx>=2.1.0",
    "pyarrow>=23.0.1",
    "pyogrio>=0.12.1",
]

[project.optional-dependencies]
example = [
    "contextily>=1.7.0",
    "matplotlib>=3.10.8",
]

[dependency-groups]
dev = [
    "ruff>=0.15.6",
    "ty>=0.0.22",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
